Remove `screenbullet.cc` and consolidate the screen type headers.
