+++ winbolo/src/bolo/screen.cc
@@ -69,10 +69,10 @@ Morrison <john@winbolo.com>          \0" */
 #include "players.h"
 #include "rubble.h"
 #include "screenbrainmap.h"
-#include "screenbullet.h"
 #include "screencalc.h"
 #include "screenlgm.h"
 #include "screentank.h"
+#include "screentypes.h"
 #include "scroll.h"
 #include "shells.h"
 #include "starts.h"
