+++ winbolo/src/bolo/screen.h
@@ -42,9 +42,9 @@
 #include "netpnb.h"
 #include "players.h"
 #include "rubble.h"
-#include "screenbullet.h"
 #include "screenlgm.h"
 #include "screentank.h"
+#include "screentypes.h"
 #include "swamp.h"
 #include "tankexp.h"
