--- winbolo/src/bolo/screenbullet.cc
@@ -1,30 +1,0 @@
-/*
- * $Id$
- *
- * Copyright (c) 1998-2008 John Morrison.
- *
- * This program is free software; you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation; either version 2 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- */
-
-/*********************************************************
- *Name:          Shells
- *Filename:      shells.c
- *Author:        John Morrison
- *Creation Date: 25/12/98
- *Last Modified: 25/12/98
- *Purpose:
- *  Responsable for Shells tracking/collision detect etc.
- *********************************************************/
-
-#include "screenbullet.h"
-
-#include "global.h"
-#include "util.h"
